requirements.txt
dbt-postgres
psycopg2-binary
numpy
sqlalchemy
pandas
requests
python-dotenv
